Added unit and 2e2 tests

+++ TableEdit/src/Domain/Order.cs
@@ -8,12 +8,12 @@ namespace Domain
         public string Name { get; set; }
         public string Description { get; set; }
         public int Count { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public double Discount { get; set; }
         public DateTime OrderDate { get; set; }
         public int DeliveryDaysCount { get; set; }
 
-        public decimal PriceWithDiscount => Price - (Price * (decimal) Discount / 100);
+        public double PriceWithDiscount => Price - (Price * Discount / 100);
 
         public DateTime DeliveryDate => OrderDate.AddDays(DeliveryDaysCount);
     }
